// Project_Apex/frontend/src/api/models/AdminKycItem.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AdminKycItem = {
    id: string;
    email: string;
    full_name?: string | null;
    kyc_status: string;
    kyc_notes?: string | null;
    kyc_submitted_at?: string | null;
    last_login_at?: string | null;
};