// Project_Apex/frontend/src/api/models/ROIExecutionPushRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ROIExecutionPushRequest = {
    trader_id?: string | null;
    user_id?: string | null;
    roi_percent: number;
    symbol: string;
    note?: string | null;
};